Split DTOs in v1 and v2 packages
Issue: IBP-3182
Reviewer: Clary
