+++ src/test/java/org/ibp/api/java/impl/middleware/germplasm/GermplasmServiceImplTest.java
@@ -4,7 +4,7 @@ package org.ibp.api.java.impl.middleware.germplasm;
 import com.google.common.collect.Lists;
 import org.generationcp.middleware.domain.germplasm.GermplasmDTO;
 import org.generationcp.middleware.domain.gms.search.GermplasmSearchParameter;
-import org.generationcp.middleware.domain.search_request.GermplasmSearchRequestDto;
+import org.generationcp.middleware.domain.search_request.v1.GermplasmSearchRequestDto;
 import org.generationcp.middleware.manager.api.GermplasmDataManager;
 import org.generationcp.middleware.manager.api.LocationDataManager;
 import org.generationcp.middleware.pojos.Germplasm;
